+++ src/main/common/utils/priorityDebts/priorityDebtsConstants.ts
