+++ src/main/views/features/response/statementOfMeans/priority-debts.njk
@@ -55,9 +55,3 @@
 </div>
 
 {% endblock %}
-
-{% block bodyEnd %}
-{# Run JavaScript at end of the <body>, to avoid blocking the initial render. #}
-{% include "webpack/js.njk" %}
-{% include "../../../webpack/income-expense-scripts.njk" %}
-{% endblock %}
