+++ src/main/views/govukTemplate.njk
@@ -132,9 +132,7 @@
 
   <script src="{{ asset_paths['js_vendor'] }}/govuk-template.js?{{ govuk_template_version }}"></script>
 
-  {% block body_end %}
-    {% block additional_scripts %}{% endblock %}  
-  {% endblock %}
+  {% block body_end %}{% endblock %}
 
   <script>if (typeof window.GOVUK === 'undefined') document.body.className = document.body.className.replace('js-enabled', '');</script>
 </body>
